Move `line_polygon` function into recognition module
This function is only used as part of text recognition, not page layout
analysis.
